gateway/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
redis==5.0.3
httpx==0.27.0
python-multipart

python-jose[cryptography]
docker
asyncpg
